cheri: tools: perf: Use uintptr_t for __rb_parent_color
Sync __rb_parent_color handling with the kernel source by using
uintptr_t for pointer tagging.

On CHERI architectures, this ensures the capability metadata and
tags for the parent node pointer are preserved during bitwise
operations for node color.

+++ tools/include/linux/rbtree.h
@@ -19,9 +19,10 @@
 
 #include <linux/kernel.h>
 #include <linux/stddef.h>
+#include <linux/types.h>
 
 struct rb_node {
-	unsigned long  __rb_parent_color;
+	uintptr_t  __rb_parent_color;
 	struct rb_node *rb_right;
 	struct rb_node *rb_left;
 } __attribute__((aligned(sizeof(long))));
@@ -40,9 +41,9 @@ struct rb_root {
 
 /* 'empty' nodes are nodes that are known not to be inserted in an rbtree */
 #define RB_EMPTY_NODE(node)  \
-	((node)->__rb_parent_color == (unsigned long)(node))
+	((node)->__rb_parent_color == (uintptr_t)(node))
 #define RB_CLEAR_NODE(node)  \
-	((node)->__rb_parent_color = (unsigned long)(node))
+	((node)->__rb_parent_color = (uintptr_t)(node))
 
 
 extern void rb_insert_color(struct rb_node *, struct rb_root *);
@@ -66,7 +67,7 @@ extern void rb_replace_node(struct rb_node *victim, struct rb_node *new,
 static inline void rb_link_node(struct rb_node *node, struct rb_node *parent,
 				struct rb_node **rb_link)
 {
-	node->__rb_parent_color = (unsigned long)parent;
+	node->__rb_parent_color = (uintptr_t)parent;
 	node->rb_left = node->rb_right = NULL;
 
 	*rb_link = node;
